Add a little fluent style also in barstyle: "strict"
Hi,
I'm using the barstyle: "strict" since in the fluent style it is hard to guess the SOC from some distance. But to have a smooth flow I've modified the strict SOC-bar a litte.
Maybe you like it, too!

This module is still a main reason for my kitchen-mirror 😄  

Take care!
Chris.